requirements.txt
Django==4.2.29
openai==2.30.0
pillow==12.1.1
python-dotenv==1.2.2
django-jazzmin==3.0.4
whitenoise==6.12.0
gunicorn==25.2.0
psycopg2-binary==2.9.11
django-ckeditor==6.7.1
